@@ -1,6 +1,6 @@
 {
   "name": "ant-virtual-table",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "description": "",
   "main": "dist/index.js",
   "directories": {
@@ -40,6 +40,9 @@
     "url": "https://github.com/ctq123/ant-virtual-table/issues"
   },
   "homepage": "https://github.com/ctq123/ant-virtual-table#readme",
+  "files": [
+    "dist"
+  ],
   "devDependencies": {
     "babel-cli": "6.26.0",
     "babel-core": "6.26.3",
@@ -50,21 +53,20 @@
     "babel-preset-react": "6.24.1",
     "babel-preset-stage-0": "6.24.1",
     "css-loader": "1.0.0",
-    "enzyme": "3.1.0",
-    "enzyme-adapter-react-16": "1.0.2",
     "html-webpack-plugin": "3.2.0",
-    "jest": "21.2.1",
-    "react": "16.6.0",
-    "react-dom": "16.6.0",
-    "react-virtualized": "9.21.0",
-    "sinon": "^7.3.2",
     "style-loader": "0.23.1",
     "webpack": "4.22.0",
     "webpack-cli": "3.1.2",
     "webpack-dev-server": "3.1.10"
   },
   "dependencies": {
-    "antd": "^3.20.0",
-    "lodash.throttle": "^4.1.1"
+    "antd": "^3.26.2",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.2",
+    "lodash.throttle": "^4.1.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "jest": "^21.2.1",
+    "sinon": "^7.3.2"
   }
 }
